+++ Documentation/markdown/ReleaseNotes.md
@@ -11,6 +11,8 @@
 
   * Adjusted the density of mesh grid lines on B-spline and NURBS surfaces ([#547](https://github.com/tumcms/Open-Infra-Platform/pull/547))
  
+  * Supporting IfcSeventhOrderPolynomialSpiral ([#552](https://github.com/tumcms/Open-Infra-Platform/pull/552))
+
   * Supporting IfcThirdOrderPolynomialSpiral ([#553](https://github.com/tumcms/Open-Infra-Platform/pull/553))
 
 ## 4.0.0 Release Notes